ASoC: pil-q6v5-mss: fix size_t warning
Corrects portability issue with 32- and 64-bit portability.
Port of (CR)

Reviewed-on: http://gerrit.mot.com/787271
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Reviewed-on: http://gerrit.mot.com/788572

+++ drivers/soc/qcom/pil-q6v5-mss.c
@@ -68,7 +68,7 @@ static void log_modem_sfr(void)
 		return;
 	}
 
-	strlcpy(pil_ssr_reason, smem_reason, min(size, sizeof(pil_ssr_reason)));
+	strlcpy(pil_ssr_reason, smem_reason, min((size_t)size, sizeof(pil_ssr_reason)));
 	pr_err("modem subsystem failure reason: %s.\n", pil_ssr_reason);
 
 	smem_reason[0] = '\0';
